fix: strict C++ consistency for IDLE/panic, tickRoot, blackboard, Any, SubTree
- tickRoot: reset root status after completed ticks in all modes (C++ bt_factory.cpp)
- SyncActionNode/StatefulActionNode: IDLE/RUNNING → panic(NewLogicError) not slog+FAILURE
- All control/decorator nodes: IDLE child → panic(NewLogicError) matching C++ LogicError
- Any: preserve uint64 as uint64 (was truncating to int64)
- Blackboard HasKey: delegate to GetEntry for parent/remapping chain lookup
- CreatePort: add IsAllowedPortName validation
- SubTree registration: NodeType.Subtree + _autoremap port

+++ core/any.go
@@ -43,7 +43,7 @@ func AnyOf(v interface{}) Any {
 	case uint32:
 		return Any{value: int64(val), originalType: reflect.TypeOf(val)}
 	case uint64:
-		return Any{value: int64(val), originalType: reflect.TypeOf(val)}
+		return Any{value: val, originalType: reflect.TypeOf(val)}
 	case float32:
 		return Any{value: float64(val), originalType: reflect.TypeOf(val)}
 	case float64:
